+++ test/config/with_symbol_convert.rb
@@ -1,0 +1,1 @@
+io_selector_backend :ruby